Refactor: move to file
